+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### [14.1.0](https://kaos.sh/ek/14.1.0)
+
+- **`[req]`** `PostFile` replaced by `SendFile` with support of different HTTP methods
+- **`[req]`** Added method `Header.Has`
+- **`[req]`** Code refactoring
+
 ### [14.0.0](https://kaos.sh/ek/14.0.0)
 
 - **`[cache/fs]`** Added method `Stop`